+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/models/AnalyzeDocumentOptions.java
@@ -8,7 +8,7 @@ import com.azure.core.annotation.Fluent;
 import java.util.List;
 
 /**
- * Options that may be passed when using analyze document API on Document Analysis client.
+ * The configurable options to be passed when using analyze document API on Document Analysis client.
  */
 @Fluent
 public final class AnalyzeDocumentOptions {