+++ sdk/communication/azure-communication-email/CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### Other Changes
 
+## 1.0.3 (2023-06-20)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-communication-common` from `1.2.8` to version `1.2.9`.
+- Upgraded `azure-core-http-netty` from `1.13.3` to version `1.13.4`.
+- Upgraded `azure-core` from `1.39.0` to version `1.40.0`.
+
 ## 1.0.2 (2023-05-23)
 
 ### Other Changes